restructured files and removed functions, found bug with updating after adding employee
