+++ employee_data.json
@@ -1,3 +1,1 @@
-[{"first name": "Sadie", "last name": "KC", "id": "1239", "doe": "May 10th 2023", "salary": "70000", "department": "HR"}, {"first name": "Omar", "last name": "Egypt", "id": "5836", "doe": "March 8th 2024", "salary": "100000", "department": "HR"}, {"first name": "Shruti", "last name": "Bara", "id": "3027", "doe": "August 31st, 2023", "salary": "90000", "department": "Sales"}, {"first name": "Steven", "last name": "Chan", "id": "2365", "doe": "April 13th, 2023", "salary": "85000", "department": "Web Dev"}, {"first name": "Sara", "last name": "woddy", "id": "25", "doe": "may", "salary": "20000", "department": "sarasdept"}, {"first name": "Chicken", "last name": "Joe", "id": "90", "doe": "june", "salary": "10000000", "department": "surfing"}]
-
-
+[{"first name": "Sadie", "last name": "KC", "id": "1239", "doe": "May 10th 2023", "salary": "70000", "department": "HR"}, {"first name": "Omar", "last name": "Egypt", "id": "5836", "doe": "March 8th 2024", "salary": "100000", "department": "HR"}, {"first name": "Shruti", "last name": "Bara", "id": "3027", "doe": "August 31st, 2023", "salary": "90000", "department": "Sales"}, {"first name": "Steven", "last name": "Chan", "id": "2365", "doe": "April 13th, 2023", "salary": "85000", "department": "Web Dev"}, {"first name": "Sara", "last name": "woddy", "id": "25", "doe": "may", "salary": "20000", "department": "sarasdept"}, {"first name": "John", "last name": "Chance the Rapper", "id": "888", "doe": "March 3rd 2018", "salary": "10000000", "department": "Hitman"}, {"first name": "Chicken", "last name": "Joe", "id": "1234", "doe": "WE dont care", "salary": "232345", "department": "Data"}]
